requirements.txt
langchain 
langchain-community 
langchain-groq 
unstructured 
selenium
chromadb
pandas
python-dotenv